# api/underfish-modified-api.openapi.yaml
openapi: 3.0.3
info:
    title: Underfish Platform API
    description: |
        Полная OpenAPI 3.0 спецификация публичных и внутренних HTTP ручек Underfish Platform,
        собранная по контроллерам authService, profileService, communityService, eventService,
        locationService, fileStorageService, gateway routes и notification endpoints из legacy underfish.

        Основной публичный вход — gateway `http://localhost:8080`. Профили публично доступны как
        `/api/v1/users/{user_id}`; gateway переписывает этот путь в profileService на
        `/api/v1/profiles/{user_id}`. Внутренние сервисные URL оставлены в `servers` для локальной
        отладки прямых сервисных вызовов.
    version: 2.0.0
servers:
    - url: http://localhost:8080
      description: Local API Gateway
    - url: https://underfish.ru
      description: Production API Gateway
    - url: http://localhost:8091
      description: Auth Service direct
    - url: http://localhost:8082
      description: Profile Service direct
    - url: http://localhost:8084
      description: Community Service direct
    - url: http://localhost:8083
      description: Event Service direct
    - url: http://localhost:8081
      description: Location Service direct
    - url: http://localhost:8085
      description: File Storage Service direct
...
paths:
...
    /api/v1/events/{event_id}/reviews:
        parameters:
            - $ref: '#/components/parameters/EventId'
        post:
            tags: [Events]
            summary: Создать отзыв о мероприятии
            operationId: createEventReview
            security: [{ bearerAuth: [] }]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ReviewRequest'
            responses:
                '201':
                    description: Отзыв создан
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ReviewResponse'
        get:
            tags: [Events]
            summary: Получить отзывы мероприятия
            operationId: listEventReviews
            responses:
                '200':
                    description: Список отзывов
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ReviewResponse'
    /api/v1/reviews/{review_id}:
        delete:
            tags: [Events]
            summary: Удалить отзыв
            operationId: deleteReview
            security: [{ bearerAuth: [] }]
            parameters:
                - $ref: '#/components/parameters/ReviewId'
            responses:
                '204': { description: Отзыв удалён }
                '404': { $ref: '#/components/responses/NotFound' }
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    parameters:
...
        EventId:
            name: event_id
            in: path
            required: true
            schema: { type: string, format: uuid }
...
        ReviewId:
            name: review_id
            in: path
            required: true
            schema: { type: string, format: uuid }
...
    responses:
...
        NotFound:
            description: Ресурс не найден
            content:
                application/json:
                    schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    schemas:
...
        ReviewRequest:
            type: object
            required: [rating, comment]
            properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string, minLength: 1 }
        ReviewResponse:
            type: object
            required: [review_id, user_id, event_id, rating, comment, created_at]
            properties:
                review_id: { type: string, format: uuid }
                user_id: { type: string, format: uuid }
                event_id: { type: string, format: uuid }
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
                created_at: { type: string, format: date-time }
...
        ErrorResponse:
            type: object
            properties:
                timestamp: { type: string, format: date-time }
                status: { type: integer }
                error: { type: string }
                message: { type: string, nullable: true }
                path: { type: string, nullable: true }